+++ check.go
@@ -7,16 +7,35 @@ import (
 
 // Check checks if the function signatures and argument types match and returns any errors
 func (q *Queue) Check() (err error) {
-	var piped []reflect.Type
-	for i, fn := range q.functions {
-		piped, err = q.validateFn(fn, i, piped)
+	return q.check(nil)
+}
+
+func (q *Queue) checkAndReturn(piped []reflect.Type) (returns []reflect.Type, err error) {
+	for i, c := range q.calls {
+
+		piped, err = q.validateFn(c, i, piped)
+
 		if err != nil {
 			return
 		}
+
+		for j, tee := range q.tees[i] {
+			_, err = q.validateFn(tee, i*100+j, piped)
+			if err != nil {
+				return
+			}
+		}
 	}
+	returns = piped
 	return
 }
 
+func (q *Queue) check(piped []reflect.Type) (err error) {
+	_, err = q.checkAndReturn(piped)
+	return
+}
+
+// validate the number of arguments
 func validateNums(fn reflect.Type, args []reflect.Type) (numIns int, numArgs int, diff int, err error) {
 	numIns = fn.NumIn()
 	numArgs = len(args)
@@ -45,6 +64,7 @@ func validateNums(fn reflect.Type, args []reflect.Type) (numIns int, numArgs int
 	return
 }
 
+// validates the arguments
 func validateArgs(fn reflect.Type, args []reflect.Type) error {
 	numIns, _, diff, err := validateNums(fn, args)
 
@@ -63,6 +83,7 @@ func validateArgs(fn reflect.Type, args []reflect.Type) error {
 	if fn.IsVariadic() {
 		limit -= 1
 	}
+
 	for i := 0; i < limit; i++ {
 		is := args[i]
 		should := fn.In(i)
@@ -91,7 +112,21 @@ func validateArgs(fn reflect.Type, args []reflect.Type) error {
 
 // validateFn validates the function at position i in the queue
 func (q *Queue) validateFn(c *call, i int, piped []reflect.Type) (returns []reflect.Type, err error) {
-	// fn := q.functions[i]
+	// fn := q.calls[i]
+
+	if c.function.Type() == queuersType {
+		qs := c.function.Interface().([]Queuer)
+
+		for _, qq := range qs {
+			piped, err = qq.Queue().checkAndReturn(piped)
+			if err != nil {
+				return
+			}
+		}
+		returns = piped
+		return
+	}
+
 	if c.function.Kind() != reflect.Func {
 		invErr := InvalidFunc{}
 		invErr.ErrorMessage = fmt.Sprintf("%#v is no func", c.function.Type().String())
@@ -109,26 +144,52 @@ func (q *Queue) validateFn(c *call, i int, piped []reflect.Type) (returns []refl
 	}
 
 	all := []reflect.Type{}
-	/*
-		args, hasArgs := q.arguments[i]
-		if hasArgs {
-			for _, p := range args {
-				if _, isPipe := p.(pipe); isPipe {
-					all = append(all, piped...)
-				} else {
-					all = append(all, reflect.TypeOf(p))
+
+	for j, p := range c.arguments {
+		switch a := p.(type) {
+		case pipe:
+			all = append(all, piped...)
+		case *call:
+			returns, err = q.validateFn(a, i*100+j*10, piped)
+			if err != nil {
+				return
+			}
+			all = append(all, returns...)
+
+		case callrun:
+			returns = piped
+			for _, qe := range a {
+				returns, err = qe.Queue().checkAndReturn(returns)
+				if err != nil {
+					return
 				}
 			}
-		}
-	*/
-	for _, p := range c.arguments {
-		if _, isPipe := p.(pipe); isPipe {
-			all = append(all, piped...)
-		} else {
+
+			all = append(all, returns...)
+
+			// TODO: all returns should match with the input arguments of the function
+			// no idea how to check it in a reasonable way (without too much overhead)
+		case callfallback:
+			for _, qe := range a {
+				returns, err = qe.Queue().checkAndReturn(piped)
+				if err != nil {
+					return
+				}
+			}
+
+			all = append(all, returns...)
+
+		default:
 			all = append(all, reflect.TypeOf(p))
 		}
 	}
 	ftype := c.function.Type()
+
+	for ia := range all {
+		if all[ia] == nil {
+			all[ia] = ftype.In(ia)
+		}
+	}
 
 	err = validateArgs(ftype, all)
 	if err != nil {
@@ -173,13 +234,53 @@ func (q *Queue) CheckAndRun() (err error) {
 	return q.Run()
 }
 
-// CheckAndFallback first runs Check() to see, if there are any type errors in the
-// function signatures or arguments and returns them. Without such errors,
-// it then calls Fallback()
-func (q *Queue) CheckAndFallback() (i int, err error) {
-	err = q.Check()
-	if err != nil {
+// TeeAndCheckAndFallback tees the given queues and in the run checks
+// them before running the Fallback method
+func (q *Queue) TeeAndCheckAndFallback(feededQs ...Queuer) *Queue {
+	fn := func(args ...interface{}) (err error) {
+		for _, qe := range feededQs {
+			err = qe.Queue().check(toTypes(args))
+			if err != nil {
+				return err
+			}
+
+		}
+
+		errHandler := q.defaultErrHandler()
+		for _, qe := range feededQs {
+			err = qe.Queue().run(toValues(args))
+			if err == nil {
+				break
+			}
+		}
+
+		if err != nil {
+			err2 := errHandler.HandleError(err)
+			q.logDebug("[E] %T(%#v) => %#v", errHandler, err, err2)
+			err = err2
+		}
 		return
 	}
-	return q.Fallback()
+	q.Tee(fn, PIPE)
+	return q
+}
+
+// TeeAndCheckAndRun tees the given queues and in the run checks
+// them before running the Run method
+func (q *Queue) TeeAndCheckAndRun(feededQs ...Queuer) *Queue {
+	fn := func(args ...interface{}) error {
+		for _, feeded := range feededQs {
+			err := feeded.Queue().check(toTypes(args))
+			if err != nil {
+				return err
+			}
+			err = feeded.Queue().run(toValues(args))
+			if err != nil {
+				return err
+			}
+		}
+		return nil
+	}
+	q.Tee(fn, PIPE)
+	return q
 }