+++ q/q.go
@@ -36,23 +36,64 @@ import (
 )
 
 var (
-	V      = queue.PIPE
-	STOP   = queue.STOP
-	IGNORE = queue.IGNORE
-	PANIC  = queue.PANIC
+	V         = queue.PIPE
+	STOP      = queue.STOP
+	IGNORE    = queue.IGNORE
+	PANIC     = queue.PANIC
+	Call      = queue.Call
+	CallNamed = queue.CallNamed
+	Get       = queue.Get
+	Set       = queue.Set
+	Collect   = queue.Collect
+	Value     = queue.Value
+	Ok        = queue.Ok
+	Fallback  = queue.Fallback
+	Run       = queue.Run
 )
 
 type (
+	add struct {
+		fn   interface{}
+		args []interface{}
+		name string
+	}
+
+	setname struct {
+		name string
+	}
+
+	tee struct {
+		fn   interface{}
+		args []interface{}
+		name string
+	}
+
+	teeRun struct {
+		qs       []queue.Queuer // []QFunc
+		validate bool
+	}
+
+	sub struct {
+		qs []queue.Queuer
+	}
+
+	teeFallback struct {
+		qs       []queue.Queuer // []QFunc
+		validate bool
+	}
+
 	run struct {
 		validate bool
 		err      error
 	}
 
-	fallback struct {
-		validate bool
-		err      error
-		pos      int
-	}
+	/*
+		fallback struct {
+			validate bool
+			err      error
+			pos      int
+		}
+	*/
 
 	log struct {
 		writer  io.Writer
@@ -63,9 +104,82 @@ type (
 		handler queue.ErrHandler
 	}
 
+	getQ struct {
+		Queue *queue.Queue
+	}
+
 	// QFunc is a function that manages a queue and returns itself for chaining
 	QFunc func(fn interface{}, params ...interface{}) QFunc
 )
+
+func (q QFunc) Queue() *queue.Queue {
+	var qq = &getQ{}
+	q(qq)
+	return qq.Queue
+}
+
+func (q QFunc) Tee(fn interface{}, args ...interface{}) QFunc {
+	var r = &tee{fn: fn, args: args}
+	q(r)
+	return q
+}
+
+func (q QFunc) SetName(name string) QFunc {
+	q.Queue().SetName(name)
+	return q
+}
+
+func (q QFunc) Name() string {
+	return q.Queue().Name()
+}
+
+func (q QFunc) TeeNamed(name string, fn interface{}, args ...interface{}) QFunc {
+	var r = &tee{fn: fn, args: args, name: name}
+	q(r)
+	return q
+}
+
+func (q QFunc) TeeAndRun(qs ...queue.Queuer) QFunc {
+	var r = &teeRun{qs: qs}
+	q(r)
+	return q
+}
+
+func (q QFunc) Sub(qs ...queue.Queuer) QFunc {
+	var r = &sub{qs: qs}
+	q(r)
+	return q
+}
+
+func (q QFunc) TeeAndCheckAndRun(qs ...queue.Queuer) QFunc {
+	var r = &teeRun{qs: qs, validate: true}
+	q(r)
+	return q
+}
+
+func (q QFunc) TeeAndFallback(qs ...queue.Queuer) QFunc {
+	var r = &teeFallback{qs: qs}
+	q(r)
+	return q
+}
+
+func (q QFunc) TeeAndCheckAndFallback(qs ...queue.Queuer) QFunc {
+	var r = &teeFallback{qs: qs, validate: true}
+	q(r)
+	return q
+}
+
+func (q QFunc) Add(fn interface{}, args ...interface{}) QFunc {
+	var r = &add{fn: fn, args: args}
+	q(r)
+	return q
+}
+
+func (q QFunc) AddNamed(name string, fn interface{}, args ...interface{}) QFunc {
+	var r = &add{fn: fn, args: args, name: name}
+	q(r)
+	return q
+}
 
 // Run runs the queue
 func (q QFunc) Run() error {
@@ -81,18 +195,6 @@ func (q QFunc) CheckAndRun() error {
 	var r = &run{validate: true}
 	q(r)
 	return r.err
-}
-
-func (q QFunc) CheckAndFallback() (int, error) {
-	var r = &fallback{validate: true}
-	q(r)
-	return r.pos, r.err
-}
-
-func (q QFunc) Fallback() (int, error) {
-	var r = &fallback{}
-	q(r)
-	return r.pos, r.err
 }
 
 func (q QFunc) LogDebugTo(w io.Writer) QFunc {
@@ -124,6 +226,8 @@ func mkQFunc(q *queue.Queue) QFunc {
 			} else {
 				v.err = q.Run()
 			}
+		case *getQ:
+			v.Queue = q
 		case *onError:
 			q.OnError(v.handler)
 		case *log:
@@ -132,12 +236,24 @@ func mkQFunc(q *queue.Queue) QFunc {
 			} else {
 				q.LogErrorsTo(v.writer)
 			}
-		case *fallback:
+		case *teeRun:
 			if v.validate {
-				v.pos, v.err = q.CheckAndFallback()
+				q.TeeAndCheckAndRun(v.qs...)
 			} else {
-				v.pos, v.err = q.Fallback()
+				q.TeeAndRun(v.qs...)
 			}
+		case *sub:
+			q.Sub(v.qs...)
+		case *teeFallback:
+			if v.validate {
+				q.TeeAndCheckAndFallback(v.qs...)
+			} else {
+				q.TeeAndFallback(v.qs...)
+			}
+		case *tee:
+			q.TeeNamed(v.name, v.fn, v.args...)
+		case *add:
+			q.AddNamed(v.name, v.fn, v.args...)
 		default:
 			q.Add(fn, i...)
 		}