+++ queue.go
@@ -6,10 +6,8 @@ import (
 )
 
 type Queue struct {
-	startValues []reflect.Value
-
-	// queue of functions
-	functions []*call
+	// queue of to be run calls
+	calls []*call
 
 	errHandler ErrHandler
 
@@ -17,31 +15,46 @@ type Queue struct {
 
 	logverbose bool
 
-	// queues into which is feeded at position int
-	feed map[int][]*Queue
-
-	// queue of functions that are piped into at a certain point
+	// queue of calls that are piped into at a certain point
 	// however their return values will be discarded (apart from errors),
 	// so they should take pointers to write something to them
 	tees map[int][]*call
 
+	subs map[int][]Queuer
+
 	// optional name of the queue (for logging and debugging)
-	Name string
+	name string
 }
 
 // New creates a new function queue
 //
-// Use Add() for adding functions to the Queue.
+// Use Add() for adding calls to the Queue.
 //
 // Use OnError() to set a custom error handler.
 //
 // The default error handler is set by the runner function, Run() or Fallback().
 //
-// Use one of these runner functions to run the queue.
+// Use one of these runner calls to run the queue.
 func New() *Queue {
 	return &Queue{
-		feed:        map[int][]*Queue{},
-		startValues: []reflect.Value{},
-		tees:        map[int][]*call{},
+		tees: map[int][]*call{},
+		subs: map[int][]Queuer{},
 	}
 }
+
+func (q *Queue) SetName(name string) *Queue {
+	q.name = name
+	return q
+}
+
+func (q *Queue) Name() string {
+	return q.name
+}
+
+func (q *Queue) Queue() *Queue { return q }
+
+type Queuer interface {
+	Queue() *Queue
+}
+
+var queuersType = reflect.TypeOf([]Queuer{})